requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apiclient==1.0.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
dnspython==2.3.0
executing==1.2.0
ffprobe==0.5
filetype==1.2.0
Flask==2.2.3
frozenlist==1.3.3
google-api-core
google-api-python-client==2.48.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-texttospeech
google-cloud-translate
google-search-results==2.4.2
googleapis-common-protos
googletrans==4.0.0rc1
grpcio==1.53.0
grpcio-status
gTTS==2.3.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.1.0
importlib-resources==5.12.0
inflection==0.5.1
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
multidict==6.0.4
nltk==3.8.1
numpy==1.24.2
openai==0.27.2
openapi==1.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.38
proto-plus
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
Pygments==2.14.0
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.3
pyspark==3.1.2
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
stack-data==0.6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
lxml==4.9.2
pytz==2023.3
autocommand==2.2.2
mongo-connector==3.1.1
mysql-connector-python
pysolr==3.9.0
nltk==3.8.1
pandas==2.0.1
Flask-APScheduler==1.12.4
flask-crontab==0.1.2
